requirements.txt
numpy
Pillow
torch>=1.7
torchvision>=0.8
tqdm
wandb
onnxruntime
albumentations
thop
